brainscore_vision/models/temporal_model_mmaction2/requirements.txt
mmaction2 @ git+https://github.com/YingtianDt/mmaction2.git@533edc3
mmengine